requirements.txt
Django==3.2.6
djangorestframework==3.12.4
flake8==3.9.2
psycopg2==2.9.1
dj-rest-auth==2.1.11
django-allauth==0.45.0
djangorestframework-camel-case==1.2.0
django-invitations==1.9.3
